Update
-BMS Houndeyes
-HL2 Beta Stalkers
-Sniper changes
-Smoke grenade for metropolice
-Combine Shield Wall for multiplayer
-Added BLOOD_COLOR_SYNTH

+++ src/game/server/hl1/hl1_npc_headcrab.h
@@ -45,7 +45,7 @@ public:
 	int RangeAttack1Conditions ( float flDot, float flDist );
 	virtual float GetDamageAmount( void );
 	virtual float GetSpecialDamageAmount(void) { return 0.0f; }
-	virtual int GetSpecialDamageType(void) { return DMG_GENERIC; }
+	virtual int64 GetSpecialDamageType(void) { return DMG_GENERIC; }
